// models/cryptoCoin.ts
export interface CryptoCoin {
  id: number;
  updatedAt: string;
  name: string;
  symbol: string;
  price: string;
  rank: number;
  description: string;
  iconUrl: string;
  websiteUrl: string;
  coinrankingUrl: string;
  marketCap: string;
  lastDayVolume: string;
  numberOfMarkets: number;
  numberOfExchanges: number;
  coinrankingId: string;
  allTimeHigh: string;
}